feat: add auth0react library, login/logout buttons
User accounts are now created and hosted on personal auth0 account. Can
be fairly straightforward change by creating a new auth0 account and
changing the environmenmt variables in production.

I currently have a copy of the .env.local file with the environment
variables to make this work locally, and can share to other developers
as needed. Just @tbn33 on the phlask slack channel or DM me directly.